build: fix angular cli commonjs warnings for polyfills
The polyfilly file was incorrectly imported through the `main` file
of the CLI project. This caused undesired CommonJS dependency
warnings for ZoneJS. We fix this by removing the import to the
polyfills and letting the CLI handle the polyfills as usually done
in fresh CLI projects.

Also removes the zone-error bundle from the devmode environment. It
will cause unnecessary CommonJS dep warnings in dev mode and also
ZoneJS generally has the potential of hiding errors (as seen with
the `Object.defineProperty` patch that was enabled by default in
non-evergreen Zone bubdles), so we should just remove it. It seems
better aligning devmode as much as possible with prodmode to catch
issues more quickly. Debugging without `zone-error` is not any
different from common Angular/ZoneJS errors as seen in the components
or framework repositories.